pyproject.toml
[project]
name = "my-prefect-server"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "mps",
    "dbt-duckdb>=1.9",
    "prefect-dbt>=0.6",
]

[tool.uv.sources]
mps = { workspace = true }

[tool.uv.workspace]
members = ["packages/*"]

[dependency-groups]
dev = [
    "ty>=0.0.23",
]

[tool.ty.src]
include = ["flows", "packages/mps/src"]
exclude = ["**/__pycache__", ".venv", ".git"]

[tool.ty.environment]
python-version = "3.14"

[tool.ty.terminal]
error-on-warning = true
